alx_travel_app/requirements.txt
amqp==5.3.1
asgiref==3.7.2
billiard==4.2.1
celery==5.5.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distlib==0.3.7
Django==4.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.10
Faker==37.4.2
filelock==3.12.2
idna==3.4
inflection==0.5.1
Jinja2==3.1.2
kombu==5.5.4
MarkupSafe==2.1.3
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==25.0
pipenv==2023.7.23
platformdirs==3.9.1
prompt_toolkit==3.0.51
pycodestyle==2.11.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2023.3
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.17.0
sqlparse==0.4.4
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.0.4
vine==5.1.0
virtualenv==20.24.2
virtualenv-clone==0.5.7
wcwidth==0.2.13
celery